refactor structure including components directory
