+++ tsconfig.json
@@ -9,7 +9,7 @@
     "baseUrl": ".",
     "paths": {
       "@config/*": ["./src/config/*"],
-      "@component/*": ["./src/component/*"]
+      "@components/*": ["./src/components/*"]
     },
     "allowJs": true,
     "skipLibCheck": true,
